smoke_tests/requirements.txt
pytest
pytest-docker-compose
requests
